test(api): handler and test case
