+++ apps/api/src/common/domainFunctions/registerHandlers.ts
@@ -15,12 +15,12 @@ type HandlerMetadata = {
 export class DomainHandlerRegistry<TDependencies> {
   #dependencies: TDependencies | null = null;
 
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  protected constructor() {}
-
-  get dependencies(): TDependencies | null {
+  get dependencies() {
     return this.#dependencies;
   }
+
+  // eslint-disable-next-line @typescript-eslint/no-empty-function
+  protected constructor() {}
 
   static domain<T>() {
     return new DomainHandlerRegistry<T>();
@@ -48,46 +48,46 @@ export class DomainHandlerRegistry<TDependencies> {
     ) as ReturnType<TResolveHandler>;
   }
 
-  prepareHandlers<THandlers extends HandlersMap<TDependencies>>(
+  prepareHandlers = <THandlers extends HandlersMap<TDependencies>>(
     handlers: THandlers,
-  ): (
+  ): ((
     dependencies: TDependencies,
-  ) => Wrap<ResolvedHandlersMap<TDependencies, THandlers>> {
+  ) => Wrap<ResolvedHandlersMap<TDependencies, THandlers>>) => {
     return (dependencies: TDependencies) => {
       this.#dependencies = dependencies;
       return this.resolveHandlers(handlers, dependencies);
     };
-  }
+  };
 
-  resolveHandlers<THandlers extends HandlersMap<TDependencies>>(
+  resolveHandlers = <THandlers extends HandlersMap<TDependencies>>(
     handlers: THandlers,
     dependencies: TDependencies,
-  ): Wrap<ResolvedHandlersMap<TDependencies, THandlers>> {
+  ): Wrap<ResolvedHandlersMap<TDependencies, THandlers>> => {
     return Object.fromEntries(
       Object.entries(handlers).map(
         ([key, fn]) => [key, this.#resolve(fn, dependencies)] as const,
       ),
     ) as Wrap<ResolvedHandlersMap<TDependencies, THandlers>>;
-  }
+  };
 
-  createHandler<Callback extends (...args: any[]) => unknown>(
+  createHandler = <Callback extends (...args: any[]) => unknown>(
     callback: (dependencies: TDependencies) => Callback,
-  ) {
+  ) => {
     Object.defineProperty(callback, $HANDLER, {
       value: {
         registry: this,
       } as HandlerMetadata,
     });
     return callback;
-  }
+  };
 
-  #resolve(
+  #resolve = (
     handler: HandlerCallback<TDependencies>,
     dependencies: TDependencies,
-  ) {
+  ) => {
     if (!Reflect.has(handler, $HANDLER)) {
       throw new Error('Given object is not a valid handler');
     }
     return handler(dependencies);
-  }
+  };
 }
