+++ apps/api/src/common/exceptions/notFoundEntityException.ts
@@ -3,7 +3,7 @@ import {HandlerError} from './handlerError';
 export class NotFoundEntityException<
   Args extends Record<string, string | number> | void = void,
 > extends HandlerError<Args> {
-  createMessage(): string {
+  createMessage(_args: Args): string {
     return 'Resource not found';
   }
 }
