+++ apps/api/src/modules/project/handlers/index.ts
@@ -1,8 +1,8 @@
 import clone from './clone';
 import createNewProject from './createNewProject';
 import findById from './findById';
-import findByUserId from './findByUserId';
+import findAllByUserId from './findByUserId';
 import update from './update';
 import updateName from './updateName';
 
-export {findById, findByUserId, createNewProject, updateName, update, clone};
+export {findById, findAllByUserId, createNewProject, updateName, update, clone};
