+++ apps/api/src/modules/project/handlers/update.ts
@@ -1,11 +1,12 @@
 import {createHandler} from '../handler';
+import {ProjectUpdateRequest, ProjectUpdateResponse} from '../schema';
 
 export default createHandler(({repository}) => {
-  return async function updateName(
+  return async function update(
     userId: string,
     projectId: string,
-    newName: string,
-  ) {
-    return repository.updateProjectName(projectId, newName);
+    data: ProjectUpdateRequest,
+  ): Promise<ProjectUpdateResponse> {
+    return repository.updateProject(userId, projectId, data);
   };
 });
