+++ apps/api/src/plugins/eventRegistry.ts
@@ -1,0 +1,18 @@
+import {FastifyPluginAsync} from 'fastify';
+import fp from 'fastify-plugin';
+import {EventRegistry} from '../common/domainFunctions/eventRegistry';
+
+declare module 'fastify' {
+  interface FastifyInstance {
+    eventRegistry: EventRegistry;
+  }
+}
+
+const eventRegistryPlugin: FastifyPluginAsync = fp(
+  async fastify => {
+    fastify.decorate('eventRegistryPlugin', new EventRegistry());
+  },
+  {name: 'eventRegistry'},
+);
+
+export default eventRegistryPlugin;
