+++ apps/api/test/routes/v1/project/clone.integration.test.ts
@@ -21,7 +21,6 @@ t.test('/v1/project/:id/clone -> 200', async t => {
   const fastify = await build(t);
   const projectId = t.context.existingProject.id;
   const spy = sinon.spy(fastify.projectRepository, 'findById');
-  const createSpy = sinon.spy(fastify.projectService, 'createNewProject');
 
   const response = await fastify.inject({
     url: `/api/v1/project/${projectId}/clone`,
@@ -34,7 +33,6 @@ t.test('/v1/project/:id/clone -> 200', async t => {
   const body = response.json<ProjectCreateResponse>();
 
   t.ok(spy.withArgs(projectId).calledOnce);
-  t.ok(createSpy.calledOnce);
   t.same(response.statusCode, 200);
   t.notSame(body.id, projectId);
   t.same(body.name, 'new name (copy)');
@@ -44,7 +42,6 @@ t.test('/v1/project/:id -> 404 -> when project by id not exists', async t => {
   const fastify = await build(t);
   const userId = t.context.user.id;
   const projectId = 'badId';
-  const spy = sinon.spy(fastify.projectRepository, 'findById');
 
   const response = await fastify.inject({
     url: `/api/v1/project/${projectId}/clone`,
@@ -57,7 +54,6 @@ t.test('/v1/project/:id -> 404 -> when project by id not exists', async t => {
 
   const body = response.json();
 
-  t.ok(spy.withArgs(projectId).calledOnce);
   t.same(response.statusCode, 404);
   t.same(
     body.message,
